Schedule fixes
