+++ frontend/src/locale/index.ts
@@ -818,4 +818,5 @@ export const messages = {
   },
 };
 
+messages["default"] = messages["en"]
 export default messages;
